requirements.txt
asgiref==3.2.10
bracex==2.1.1
casbin==1.3.0
Django==3.1.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Pillow==8.3.1
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2021.1
simpleeval==0.9.10
sqlparse==0.4.1
wcmatch==8.2
